Exported FeatureSegmentsManager interface to browser-api module
