+++ app/src/main/java/com/duckduckgo/app/onboarding/di/WelcomePageModule.kt
@@ -17,13 +17,13 @@
 package com.duckduckgo.app.onboarding.di
 
 import android.content.Context
-import com.duckduckgo.app.featureusage.FeatureSegmentsManager
 import com.duckduckgo.app.global.DefaultRoleBrowserDialog
 import com.duckduckgo.app.global.RealDefaultRoleBrowserDialog
 import com.duckduckgo.app.global.install.AppInstallStore
 import com.duckduckgo.app.onboarding.ui.page.WelcomePageViewModelFactory
 import com.duckduckgo.app.statistics.pixels.Pixel
 import com.duckduckgo.appbuildconfig.api.AppBuildConfig
+import com.duckduckgo.browser.api.featureusage.FeatureSegmentsManager
 import dagger.Module
 import dagger.Provides
